Fix pr comments

+++ src/controllers/PostController.js
@@ -1,7 +1,10 @@
 
+const SectionService = require('../services/SectionService');
 const PostService = require('../services/PostService');
 const { serializePost } = require('../utils/serializers');
 
+const ForbiddenError = require('../constants/errors/ForbiddenError');
+const InvalidFieldError = require('../constants/errors/InvalidFieldError');
 
 const create = async (req, res) => {
     const {
@@ -11,9 +14,17 @@ const create = async (req, res) => {
         endDate,
         type,
         section,
+        attachments,
     } = req.body;
 
-    const attachments = [];
+    if (type === 'Event' && !(new Date(endDate)).getDate()) {
+        throw new InvalidFieldError('Fecha de finalización no es válida');
+    }
+
+    const students = await SectionService.findSectionsStudents(section);
+    if (!students.includes(req.user.id)) {
+        throw new ForbiddenError('Usted no es parte de esta sección');
+    }
 
     const post = await PostService.create({
         title,
@@ -32,6 +43,7 @@ const create = async (req, res) => {
 };
 
 const update = async (req, res) => {
+    const { id } = req.params;
     const {
         title,
         description,
@@ -39,11 +51,20 @@ const update = async (req, res) => {
         endDate,
         type,
         section,
+        attachments,
     } = req.body;
 
-    const attachments = [{ name: 'x', fileURL: '' }];
+    if (type === 'Event' && !(new Date(endDate)).getDate()) {
+        throw new InvalidFieldError('Fecha de finalización no es válida');
+    }
 
-    const post = await PostService.update({
+    const post = await PostService.findById(id);
+    if (post.author !== req.user.id) {
+        throw new ForbiddenError('Este post no le pertenece');
+    }
+
+    const updated = await PostService.update({
+        id,
         title,
         description,
         startDate,
@@ -55,7 +76,7 @@ const update = async (req, res) => {
     });
     res.json({
         success: true,
-        data: post,
+        data: updated,
     });
 };
 
@@ -70,6 +91,11 @@ const details = async (req, res) => {
 
 const deletePost = async (req, res) => {
     const { id } = req.params;
+
+    const post = await PostService.findById(id);
+    if (post.author !== req.user.id) {
+        throw new ForbiddenError('Este post no le pertenece');
+    }
 
     const result = await PostService.deletePost(id);
     const success = result.ok > 0 && result.n > 0;
